requirements.txt
chromadb
fastapi
uvicorn
rich
llama-cpp-python
watchdog
Pillow
